+++ packages/amplify-frontend-javascript/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-javascript@3.3.4...amplify-frontend-javascript@3.3.5) (2021-11-26)
+
+**Note:** Version bump only for package amplify-frontend-javascript
+
+
+
+
+
 ## [3.3.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-javascript@3.3.3...amplify-frontend-javascript@3.3.4) (2021-11-23)
 
 **Note:** Version bump only for package amplify-frontend-javascript